Include MIT license text for sharded-slab in bench.rs
Since bench.rs is derived from the bench.rs in sharded-slab, and the MIT
license says “The above copyright notice and this permission notice
shall be included in all copies or substantial portions of the
Software,” we need to include the original license text.

Using a source comment to include the license text should be adequate,
since it’s unlikely anyone would want to distribute the compiled
benchmark executable.