add benchmark license text and remove from distribution
Since `bench.rs` is derived from the `bench.rs` in [`sharded-slab`](https://github.com/hawkw/sharded-slab), and the MIT license says “The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software,” we need to include the original license text.

Using a source comment to include the license text should be adequate, since it’s unlikely anyone would want to distribute the compiled benchmark executable.

----

As a follow-up, this PR excludes the benchmarks (along with a few other unnecessary files) from published crates. This makes the crate archives slightly smaller, and makes it very clear that nothing derived from `sharded-slab` is compiled into the crate library, which makes downstream license analysis easier.